change rng generation method to be more cross platform

+++ src/hash.c
@@ -1,7 +1,7 @@
 
-unsigned long hash(const char *str)
+unsigned int hash(const char *str)
 {
-	unsigned long hash = 5381;
+	unsigned int hash = 5381;
 	int c;
 
 	while ((c = *str++))
