+++ src/hash.h
@@ -1,6 +1,6 @@
 #ifndef HASH_H
 #define HASH_H
 
-unsigned long hash(const char *str);
+unsigned int hash(const char *str);
 
 #endif /* HASH_H */
